@@ -1,15 +1,16 @@
-package com.example.nathan.loadtracker.activities
+package com.example.nathan.loadtracker.ui.activities
 
 import android.content.Context
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import android.os.Bundle
-import android.support.design.widget.Snackbar
+import com.google.android.material.snackbar.Snackbar
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.example.nathan.loadtracker.R
-import com.example.nathan.loadtracker.database
-import com.example.nathan.loadtracker.models.Load
+import com.example.nathan.loadtracker.core.database.LoadTrackerDatabase
+import com.example.nathan.loadtracker.core.database.entities.JobSession
+import com.example.nathan.loadtracker.core.database.entities.Load
 import kotlinx.android.synthetic.main.fragment_load_tracking.*
 import java.text.SimpleDateFormat
 import java.util.*
@@ -17,14 +18,14 @@ import java.util.*
 class TrackLoadFragment : Fragment() {
 
     private lateinit var sessionTitle: String
-    private lateinit var js: MutableList<Load>
+    private lateinit var js: JobSession
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         sessionTitle = activity?.title?.toString()!!
 
-        js = context?.database?.getLoadsForSession(sessionTitle)!!.toMutableList()
+        js = LoadTrackerDatabase.getJobSession(sessionTitle)
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
@@ -56,29 +57,28 @@ class TrackLoadFragment : Fragment() {
             }
 
             val c = Calendar.getInstance()
-            context?.database?.addLoad(
-                    Load(
-                            id = null,
-                            title = sessionTitle,
-                            driver = driverNameInput.text.toString(),
-                            unitId = unitIDInput.text.toString(),
-                            material = materialInput.text.toString(),
-                            timeLoaded = SimpleDateFormat("HH:mm:ss.SSS").format(c.time),
-                            dateLoaded = SimpleDateFormat("yyyy/MM/dd").format(c.time),
-                            created = SimpleDateFormat("yyyy/MM/dd").format(c.time),
-                            modified = null,
-                            companyName = companyNameInput.text.toString()
-                    )
-            )
+
+            js.loads.add(Load(
+                    driver = driverNameInput.text.toString(),
+                    unitId = unitIDInput.text.toString(),
+                    material = materialInput.text.toString(),
+                    timeLoaded = SimpleDateFormat("HH:mm:ss.SSS").format(c.time),
+                    dateLoaded = SimpleDateFormat("yyyy/MM/dd").format(c.time),
+                    created = SimpleDateFormat("yyyy/MM/dd").format(c.time),
+                    modified = null,
+                    companyName = companyNameInput.text.toString()
+            ))
 
             Snackbar.make(view, "Tracked!", Snackbar.LENGTH_LONG).show()
         }
 
-        if (js.isNotEmpty()) {
-            materialInput.setText(js[js.size - 1].material)
-            unitIDInput.setText(js[js.size - 1].unitId)
-            driverNameInput.setText(js[js.size - 1].driver)
-            companyNameInput.setText(js[js.size - 1].companyName)
+        if (js.loads.isNotEmpty()) {
+            js.loads.let {
+                materialInput.setText(it[it.size - 1].material)
+                unitIDInput.setText(it[it.size - 1].unitId)
+                driverNameInput.setText(it[it.size - 1].driver)
+                companyNameInput.setText(it[it.size - 1].companyName)
+            }
         } else {
             val sharedPrefs = activity?.getSharedPreferences("com.example.nathan.loadtracker", Context.MODE_PRIVATE)
             driverNameInput.setText(sharedPrefs?.getString("name", ""))
