@@ -1,15 +1,15 @@
-package com.example.nathan.loadtracker.activities
+package com.example.nathan.loadtracker.ui.activities
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v7.widget.LinearLayoutManager
+import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.LinearLayoutManager
 import android.view.*
 import kotlinx.android.synthetic.main.fragment_tracking_history.*
 
 import com.example.nathan.loadtracker.R
-import com.example.nathan.loadtracker.arrayadapters.TrackingHistoryAdapter
-import com.example.nathan.loadtracker.database
-import com.example.nathan.loadtracker.models.Load
+import com.example.nathan.loadtracker.core.database.LoadTrackerDatabase
+import com.example.nathan.loadtracker.ui.arrayadapters.TrackingHistoryAdapter
+import com.example.nathan.loadtracker.core.database.entities.Load
 
 class TrackingHistoryFragment : Fragment() {
 
@@ -20,7 +20,7 @@ class TrackingHistoryFragment : Fragment() {
         super.onCreate(savedInstanceState)
 
         sessionTitle = activity?.title?.toString()!!
-        loads = context?.database?.getLoadsForSession(sessionTitle) as ArrayList<Load>
+        loads = LoadTrackerDatabase.getLoadsForSession(sessionTitle) as ArrayList<Load>
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
